+++ CHANGELOG.md
@@ -14,3 +14,10 @@
 ### Version 1.4.0
 
 - General Style and changed Extension name and branding
+
+### Version 1.5.0
+
+- Hide youtube Shorts everywhere
+- Toggle for hide Shorts in Search results
+- Changed extension name
+- Fixed badge inizialization